split fsspec tests into reading and writing
